refactor: remove BaseElement
